requirements.txt
requests
matterhook
python-telegram-bot
discord-webhook